requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
dateparser==1.2.2
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
pycryptodome==3.23.0
python-binance==1.0.32
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
six==1.17.0
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
